refactor: 도메인 수정 및 termination 수정
