+++ manager/src/main/java/com/zupzup/untact/model/auth/ManagerUserDetails.java
@@ -1,7 +1,6 @@
 package com.zupzup.untact.model.auth;
 
 import com.zupzup.untact.model.Manager;
-import com.zupzup.untact.model.Member;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
