+++ manager/src/main/java/com/zupzup/untact/service/impl/TerminationServiceImpl.java
@@ -1,12 +1,12 @@
 package com.zupzup.untact.service.impl;
 
-import com.zupzup.untact.domain.auth.seller.Seller;
-import com.zupzup.untact.domain.auth.user.User;
-import com.zupzup.untact.domain.enums.EnterState;
-import com.zupzup.untact.domain.store.Store;
 import com.zupzup.untact.exception.member.MemberException;
 import com.zupzup.untact.exception.store.StoreException;
 import com.zupzup.untact.model.Member;
+import com.zupzup.untact.model.domain.auth.seller.Seller;
+import com.zupzup.untact.model.domain.auth.user.User;
+import com.zupzup.untact.model.domain.enums.EnterState;
+import com.zupzup.untact.model.domain.store.Store;
 import com.zupzup.untact.model.dto.response.DeleteStoreListRes;
 import com.zupzup.untact.model.dto.response.StoreRes;
 import com.zupzup.untact.repository.*;
@@ -20,8 +20,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static com.zupzup.untact.exception.member.MemberExceptionType.NOT_FOUND_MEMBER;
 import static com.zupzup.untact.exception.store.StoreExceptionType.NO_MATCH_STORE;
@@ -46,21 +46,11 @@ public class TerminationServiceImpl implements TerminationService {
     public List<DeleteStoreListRes> gatherDeletionRequests() {
 
         // enterState 가 delete 인 가게 리스트
-        List<Store> sList = storeRepository.findByEnterState(EnterState.DELETE);
+        List<Store> storeList = storeRepository.findByEnterState(EnterState.DELETE);
 
-        if (sList.size() == 0) {
-
-            return new ArrayList<>();
-        }
-
-        List<DeleteStoreListRes> resList = new ArrayList<>();
-
-        for (Store s : sList) {
-
-            resList.add(modelMapper.map(s, DeleteStoreListRes.class));
-        }
-
-        return resList;
+        return storeList.stream()
+                .map(entity -> modelMapper.map(entity, DeleteStoreListRes.class))
+                .collect(Collectors.toList());
     }
 
     /**
@@ -69,14 +59,14 @@ public class TerminationServiceImpl implements TerminationService {
     @Override
     public StoreRes deletionDetail(Long id) {
 
-        Store s = storeRepository.findById(id)
+        Store store = storeRepository.findById(id)
                 .orElseThrow(() -> new StoreException(NO_MATCH_STORE));
 
         // 판매자 로그인 아이디 가져오기
-        Seller seller = sellerRepository.findById(s.getSellerId())
+        Seller seller = sellerRepository.findById(store.getSellerId())
                 .orElseThrow(() -> new MemberException(NOT_FOUND_MEMBER));
 
-        StoreRes rs = modelMapper.map(s, StoreRes.class);
+        StoreRes rs = modelMapper.map(store, StoreRes.class);
         rs.setSellerLoginId(seller.getLoginId());
 
         return rs;
@@ -88,21 +78,21 @@ public class TerminationServiceImpl implements TerminationService {
     @Override
     public String deleteToConfirm(Long id) {
 
-        Store s = storeRepository.findById(id)
+        Store store = storeRepository.findById(id)
                 .orElseThrow(() -> new StoreException(NO_MATCH_STORE));
 
         // 노출 대기 상태로 변경
-        s.setEnterState(EnterState.WAIT);
+        store.setEnterState(EnterState.WAIT);
         // 노출 대기된 시간 업데이트
-        s.setWaitStatusTimestamp(timeSetter());
+        store.setWaitStatusTimestamp(timeSetter());
 
         // 회원 탈퇴 요청 철회
-        Long sellerId = s.getSellerId();
+        Long sellerId = store.getSellerId();
         Seller seller = sellerRepository.findById(sellerId)
                         .orElseThrow(()-> new MemberException(NOT_FOUND_MEMBER));
         seller.setWantDeletion(false);
 
-        storeRepository.save(s);
+        storeRepository.save(store);
         sellerRepository.save(seller);
 
         return "Enter state is changed into WAIT";
