+++ member/src/main/resources/application.yml
@@ -10,3 +10,6 @@ spring:
       driver-class-name: org.h2.Driver
       username: sa
       password:
+
+server:
+  port: 8085
